+++ src/Error/index.js
@@ -1,14 +1,14 @@
 const ExtendableError = require('./ExtendableError');
 
 class WatchError extends ExtendableError {}
-class WatchTimeoutError extends ExtendableError {}
 class AlreadyInitializedError extends ExtendableError {}
+class NotInitializedError extends ExtendableError {}
 class InvalidDataError extends ExtendableError {}
 
 
 module.exports = {
     WatchError,
-    WatchTimeoutError,
     AlreadyInitializedError,
+    NotInitializedError,
     InvalidDataError
 };